configuration of react-navigation
