+++ package.json
@@ -13,6 +13,7 @@
     "@react-navigation/elements": "^2.2.5",
     "@react-navigation/native": "^7.0.14",
     "@react-navigation/native-stack": "^7.2.0",
+    "@tanstack/react-query": "^5.65.1",
     "react": "18.3.1",
     "react-native": "0.77.0",
     "react-native-safe-area-context": "^5.1.0",
@@ -33,6 +34,8 @@
     "@types/react": "^18.2.6",
     "@types/react-test-renderer": "^18.0.0",
     "eslint": "^8.19.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-prettier": "^5.2.3",
     "jest": "^29.6.3",
     "prettier": "2.8.8",
     "react-test-renderer": "18.3.1",
