+++ src/screens/ImageScreen.tsx
@@ -1,0 +1,11 @@
+import {Text, View} from 'react-native';
+
+function ImageScreen() {
+  return (
+    <View>
+      <Text>some text</Text>
+    </View>
+  );
+}
+
+export default ImageScreen;
