+++ src/screens/LikesScreen.tsx
@@ -1,0 +1,11 @@
+import {Text, View} from 'react-native';
+
+function LikesScreen() {
+  return (
+    <View>
+      <Text>some text</Text>
+    </View>
+  );
+}
+
+export default LikesScreen;
